Part_1/src/sw2/text_rules/requirements.txt
adal==1.2.2
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
Click==7.0
cryptography==2.8
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msrest==0.6.10
msrestazure==0.6.2
numpy==1.17.4
oauthlib==3.1.0
pandas==0.25.3
pycparser==2.19
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
SQLAlchemy==1.3.11
urllib3==1.25.7
Werkzeug==0.16.0
WTForms==2.2.1
psycopg2-binary==2.7.5
